Implement border decorations with gaps Refactor out some common test utiliity methods Update the decorations test to show borders on a simple "photo" Observe (and show in the decoration test) that shadows are not implemented at all
